+++ src/tasks/differences/file/content/enums.ts
@@ -2,7 +2,7 @@ import { Enum, EnumID, Reader } from "../../../../utils/cache2";
 import { CompareFn } from "../../differences.types";
 import { getFileDifferences } from "../file.utils";
 
-const compareEnums: CompareFn = ({ oldFile, newFile }) => {
+const compareEnums: CompareFn = async ({ oldFile, newFile }) => {
   const oldEntry = oldFile
     ? Enum.decode(
         new Reader(oldFile.file.data, {