+++ src/tasks/infoboxGenernator/infoboxes/item/item.ts
@@ -78,7 +78,7 @@ export const buildItemInfobox = async (item: Item, writeFiles = true) => {
 
     let infoboxBonuses = undefined;
     let combatStyles = undefined;
-    if (item.wearpos1 > 0) {
+    if (item.wearpos1 >= 0) {
       infoboxBonuses = new InfoboxTemplate("bonuses", {
         astab: item.params.has(STAB_ATTACK_PARAM)
           ? formatBonus(item.params.get(STAB_ATTACK_PARAM))
@@ -136,8 +136,12 @@ export const buildItemInfobox = async (item: Item, writeFiles = true) => {
             ? item.params.get(ATTACK_RANGE_PARAM) ?? "0"
             : undefined,
         combatstyle: getWeaponCategory(item),
-        image: new MediaWikiFile(`${item.name} equipped male.png`),
-        altimage: new MediaWikiFile(`${item.name} equipped female.png`),
+        ...(item.wearpos1 !== WearPos.Ammo && item.wearpos1 !== WearPos.Ring
+          ? {
+              image: new MediaWikiFile(`${item.name} equipped male.png`),
+              altimage: new MediaWikiFile(`${item.name} equipped female.png`),
+            }
+          : {}),
       });
 
       if (item.wearpos1 === WearPos.Weapon && item.category > -1) {