+++ src/utils/cache2/loaders/DBRow.ts
@@ -3,7 +3,13 @@ import { Loadable, PerFileLoadable } from "../Loadable";
 import { Reader } from "../Reader";
 import { Typed } from "../reflect";
 import { BaseVarType, ScriptVarType } from "../ScriptVarType";
-import { DBColumnID, DBRowID, DBTableID, ScriptVarID } from "../types";
+import {
+  DBColumnID,
+  DBRowID,
+  DBTableID,
+  GameValType,
+  ScriptVarID,
+} from "../types";
 
 function readTypes(r: Reader): ScriptVarID[] {
   const size = r.u8();
@@ -56,10 +62,12 @@ export class DBRow extends PerFileLoadable {
 
   public static readonly index = 2;
   public static readonly archive = 38;
+  public static readonly gameval = GameValType.DBRows;
 
   public table = <DBTableID>-1;
   public values: (string | number | bigint | undefined)[][] = [];
   public types: (ScriptVarID | undefined)[][] = [];
+  public gameVal?: string;
 
   public static decode(r: Reader, id: DBRowID): DBRow {
     const v = new DBRow(id);
@@ -100,9 +108,11 @@ export class DBTable extends PerFileLoadable {
 
   public static readonly index = 2;
   public static readonly archive = 39;
+  public static readonly gameval = GameValType.DBTables;
 
   public types: (ScriptVarID | undefined)[][] = [];
   public defaultValues: (string | number | bigint | undefined)[][] = [];
+  public gameVal?: string;
 
   public static decode(r: Reader, id: DBTableID): DBTable {
     const v = new DBTable(id);