+++ src/utils/cache2/loaders/Sprite.ts
@@ -77,6 +77,7 @@ export class Sprites extends NamedPerArchiveLoadable {
   }
 
   static readonly index = 8;
+  static readonly gameval = types.GameValType.Sprites;
 
   width!: number;
   height!: number;
@@ -84,6 +85,7 @@ export class Sprites extends NamedPerArchiveLoadable {
   sprites: Sprite[];
 
   palette!: types.PrimitiveArray<types.RGB, Uint32Array>;
+  public gameVal?: string;
 
   public static decode(r: Reader, id: types.SpriteID): Sprites {
     let chunkOffset = r.length - 2;